Fix a race condition in the startup of threads to perform parallel Karatsuba in arithlib.hpp. This showed up leading to failured on a Macbook while the same tests had run happily on Linux-x86_64, Cygwin64, armv7l and aarch64. An illustration that races in threaded code can be slippery and onlyt show up either occasionally or in a seriously platform-specific way!
git-svn-id: https://svn.code.sf.net/p/reduce-algebra/code/trunk@5143 2bfe0521-f11c-4a00-b80e-6202646ff360

+++ csl/cslbase/arithtest.cpp
@@ -496,7 +496,7 @@ int main(int argc, char *argv[])
         a = fudge_distribution_bignum(a, (int)r & 0xf);
         r = (r >> 4)%800;
         Bignum c1 = a << r;
-        Bignum p = pow(Bignum(2), (int)r);
+        Bignum p = pow(Bignum(2), (int64_t)r);
         Bignum c2 = a * p;
         Bignum c3 = a >> r;
         Bignum w = a & ~(p-1_Z);
